# backend/plugin/openapi.yaml
servers:
  - url: PLUGIN_HOSTNAME
info:
  description: Scheduling and keeping track of task status. Each time you do an action, register it.
  title: TOP Coordinator
  version: "1.0"
openapi: 3.0.1
paths:
  /coordination/view_jobs:
    get:
      summary: List the TOP Jobs that are avaialble.
      operationId: listAvailableTOPJobs
      responses:
        default:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetJobsResponse"

  /coordination/execute_job:
    post:
      summary: Used to begin doing a job. This will return the instructions for how to accomplish the job and where to store data. Just need to pass in the job_id.
      operationId: executeTOPJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobExecutionRequest"

  /coordination/log_results:
    post:
      summary: While executing the job, once a section is complete send the results back as a JSON object (result).
      operationId: logJobResults
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCompletionResults"

components:
  schemas:
    GetJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
    Job:
      type: object
      properties:
        job_id:
          type: string
        job_name:
          type: string
        job_description:
          type: string
    JobExecutionRequest:
      type: object
      properties:
        job_id:
          type: string
    JobCompletionResults:
      type: object
      properties:
        job_id:
          type: string
        results:
          type: object
          properties:
            result:
              type: object
              description: The result of this part of the job. This can be anything, but it must be a JSON object.